# routes/swagger.yaml
...
swagger: '2.0'
info:
  title: KipsWarehouse API
  description: Move your app forward with the KipsWarehouse API
  version: "1.0.0"
# the domain of the service
host: kipswarehouse.com
...
schemes:
  - https
securityDefinitions:
  JWT:
    type: apiKey
    name: Authorization
    in: header
basePath: /
produces:
  - application/json
paths:
...
  /api/users:
    get:
      security:
        - JWT: []
      description: |
        Get users' information.
      parameters:
        - name: userId
          in: query
          description: Get the information of the user with {userId}
          required: false
          type: string
      tags:
        - User
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithArrayDataType'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /api/users/apikey:
    get:
      security:
        - JWT: []
      description: |
        Get the apikey
      tags:
        - User
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithObjectDataType'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /api/users/{username}/{password}:
    post:
      security:
        - JWT: []
      description: |
        Create a new user.
      parameters:
        - name: username
          in: path
          description: username
          required: true
          type: string
        - name: password
          in: path
          description: password
          required: true
          type: string
      tags:
        - User
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithObjectDataType'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /api/users/{userId}:
    put:
      security:
        - JWT: []
      description: |
        Change an existing user's information
      parameters:
        - name: userId
          in: path
          description: id of the user to be changed.
          required: true
          type: string
        - name: displayName
          in: query
          description: displayName
          required: false
          type: string
        - name: email
          in: query
          description: email
          required: false
          type: string
        - name: permission
          in: query
          description: permission
          required: false
          type: string
      tags:
        - User
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithObjectDataType'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  ResponseWithArrayDataType:
    type: object
    properties:
      status:
        description: "success"
        type: string
      data:
        description: An array of objects
        type: array
        items:
          type: object
  ResponseWithObjectDataType:
    type: object
    properties:
      status:
        description: "success"
        type: string
      data:
        description: An object
        type: object
  Error:
    type: object
    properties:
      status:
        description: "error"
        type: string
      error:
        description: Error object
        type: object